link to "Cost and usage update" + updated figures
Hello content owners! I recently worked with the PG on this ICM https://portal.microsofticm.com/imp/v3/incidents/details/199366797/home. The PG informed us that "Azure Resource Providers are currently allowed to report usage up to 9 days after the time of usage." 

In addition to that, Support team in case #120063025002019 stated that while the period is not closed cost and usage data in ACM should be treated as preliminary. The source of truth is after the period is closed. Similar information is posted here: https://docs.microsoft.com/en-us/azure/cost-management-billing/costs/understand-cost-mgt-data#cost-and-usage-data-updates-and-retention

Due to this, budget alerts might be delayed up to several days (in worst case scenario till end of period). I would suggest to add the above link to the article about Azure Budgets, so that our customers and partners understand this specifics and we avoid further confusion. If there are additional articles on Azure Budgets, it might be worth to add this link there as well.