Update az aks create examples
When I ran the az aks create example as written, I received the following error:

```text
An RSA key file or key value must be supplied to SSH Key Value. You can use --generate-ssh-keys to let CLI generate one for you
```

However, I eventually want to use Azure AD integration and thought this might be a better solution.

The feature `--enable-aad` seemed to work in the Azure CLI 2.9.1,  But I am wondering if the feature still needs the preview mentioned on my reference page.  

In any case, I was hoping this might be a fix to an issue my error. If it is too soon because the feature is in preview, then I do understand.