Add "global"
The Ignite Book of News at https://news.microsoft.com/ignite-2020-book-of-news/#181-azure-networking-enhancements-include-cisco-sd-wan-with-azure-virtual-and-global-load-balancer-feature- calls this feature the "Global load balancing" feature.  I don't know where we are using what terms, but we need to make sure customers can find what we're announcing quickly.  Right now, it's basically impossible for a customer to get from the announcement to the documentation, because we use different terms.