missing code to read table from sql pool
The existing code does not read from SQL pool as suggested in the document and creates a reader object and not a dataframe. Adding sqlanalaytics generates a read from SQL Pool into a dataframe. It is unclear what the option is doing and I am fairly certain it is not requred.