server 2016 LTSC with v1607 missing some feature
Hi John,
Many customer doesn't take sub version in the Windows Server 2016, they are looking as supported OS 2016 level. not the sub level v1607 or v1803. I discussed with Ravi Vennapusa and opened this PR to update
FYI: many customer on Windows server LSTC branch. which supported with v1607. They never be on Windows Server v1803. which is not LSTC, which is on Semi-annual phase plan.