Change comment syntax and add backticks
The backticks were missing from the `Join-AzStorageAccountForAuth` cmdlet, leading the account to be created with inappropriate defaults. Even if there were backticks, they would have been ineffective due to the comment syntax `#` instead of `<# #>`. 

(Technically, the last comment (line 72) doesn't need the other syntax, but I did it for consistency.)