Update query-json-files.md
Before my fix, the document said "This query will return each JSON document as a separate row of the result set.", but the above code includes URL to a document which only includes a single JSON document. Therefore, this is not accurate. There is only one JSON document in the source file. The JSON document includes an array of objects and each object is a record which should be converted to a row in the table.

I changed the text to: "The JSON document in the above demo includes an array of objects. This query will return each object as a separate row in the result set."