Updated flow
Now there're 2 different options when creating a new notebook, loading to dataframe was similar to the previous example in this chapter. Also added a bit more clarification for step 7 in case someone gets confused about where is the .parquet file inside since the notebook tab is opened in that moment and not the filesystem tab.