Address GitHub issue  #63955
Input from Geert for the changes are as follows -
----------
I believe the document describes the Linux default behavior of the domain been set to “localdomain”, and how to change that to “defaultv4iddomain.com" already.

It says:
Root mapping defaults to the nobody user because the NFSv4 domain is set to localdomain.

Perhaps change that to (to clarify):
Root mapping defaults to the nobody user because the NFSv4 domain is set to localdomain by default.

Then it goes on saying:
As the above example shows, the user for file1 should be root, but it maps to nobody by default. This article shows you how to set the file1 user to root.

Perhaps change that to (to clarify):
As the above example shows, the user for file1 should be root, but it maps to nobody by default. This article shows you how to set the file1 user to root by changing the idmap Domain setting to defaultv4iddomain.com.
-----------