Update how-to-connect-install-existing-tenant.md
Removing an incorrect statement.  The values are not overwritten initially when the match occurs, but on the following sync cycle AADConnect imports the data from AAD (including the attributes containing values) and determines that these attributes do not have any values on-prem, so it Exports the null values from on-prem to AAD to put the user "in-sync". The SoA of the object is on-premises so all values must come from onpremises. If the documentation was correct then the user would become "inconsistent" with the SoA (onprem AD).
Please check the internal discussion on Teams for more details:
https://teams.microsoft.com/l/message/19:6443acc4b72c471bac9e2ba039fa713a@thread.skype/1602757841750?tenantId=72f988bf-86f1-41af-91ab-2d7cd011db47&groupId=70b122d3-4d70-4357-b412-a8051556c3a4&parentMessageId=1602757841750&teamName=CSS%20Azure%20Identity%20-%20User%20Provisioning%20%2F%20Health&channelName=AAD%20Connect&createdTime=1602757841750