Update application-gateway-waf-configuration.md
Changing 1-KB and 128-KB to 1KB and 128KB in the below. The '-' in there is misleading and causes the reader to read that as a range when it is not.

"The maximum request body size field is specified in kilobytes and controls overall request size limit excluding any file uploads. This field can range from 1-KB minimum to 128-KB maximum value. The default value for request body size is 128 KB"