Delete UPD recommendation
The document originally names two recommended solutions for profile management as "the" recommendation - FSLogix and UPD.  It could be argued that FSLogix is a variation on the UPD theme, but that's not the position in the document - instead, UPD and FSLogix are presented as two distinct technologies.  This PR removes the conflict between the recommendations.