(AzureCXP) fixes MicrosoftDocs/azure-docs/issues/65092
According to internal discussion on this there is no reference on the documentation regarding to the fact that Local Cache does not seem to work for Free and Shared SKUs. Adding " Local Cache is not supported in the **F1** or **D1** tier. " as note in Enabling Local Cache in App Service section.