(AzureCXP) fixes #65862
Private offers (or more accurately: private plans) are available for gov cloud like any other offer. A plan can also shift status from private to public and vice versa. The control on which cloud the offer will be available is in the partner’s hands and can be configured via Partner Center.