Update new-relic-limited-release-tutorial.md
Update doc to reflect that the app has left limited release. As part of this, setting up the New Relic side is now self-serve. No need to call New Relic to do that part.

Some of the Microsoft side instructions looked out of date so changed to match the current UI.