remove article. split up article
Removing https://docs.microsoft.com/en-us/azure/azure-monitor/platform/stream-monitoring-data-event-hubs#partner-tools-with-azure-monitor-integration so putting parts of it into this article first.  @anirudhcavale and  @nkiest to check the wording and for awareness. If we have a link in the portal pointing to this previous article, it should point here to this article and bookmark now.  @bwren for awareness to clear deletion of this event hub article.