Update Connection to Site-to-Site Connection and also scenario of having more than one Site-to-Site Connection
I am proposing this change because it was not clear in the document to only select one Site-to-Site Connection at a time for returning correct health results under the VPN Troubleshoot Diagnostics Tool. It caused confusion to my customer who had selected more than two site-to-site connections at a time, also he felt that this information was missing in the public document that caused failed results on one of the connections, though both connections are showing connected under Site-to-Site Connections Page on the Azure Portal. The Customer recommended having this information to select only one Site-to-Site Connection at a time on VPN Troubleshoot, be updated in the Public documentation for the VPN Troubleshoot.