correct supported type values for parameters
In visual studio code linter, supported values on the parameterization template are different from what is listed in this article. Also, when running a deployment with triggerSuffix of "number" - an error is thrown, stating that number is not a supported type. My guess is someone was using number like in typescript and this is likely just a typo, or perhaps something has changed recently and this is out of date. I updated the list of supported values and corrected anywhere using triggerSuffix number with int instead.