Add AAS logs information
Acs Authentication Service (AAS) is going to emit logs with operational information. This PR adds changes in public documentation. Most of the items are common to the other services already emitting these logs, except two specific items:

Identity - identity is a definite descriptor of a user. If an operation related to an identity is performed, the identity is logged.
Scopes - some operations are only allowed for certain scopes. These scopes are logged if relevant.
-------
cc: @mikben