Creating a "New" Hub Virtual Network using AZFW Manager ignores the Availability Zone selection
Despite selecting two or more AZs at creation time, post deployment, the properties of AZFW inside of the AzureFirewallSubnet of the new Hub virtual network points to an absence of AZs.  Furthermore, the deployment summary inside of the Azure Portal doesn't specify that AZs were selected.  This needs to be fixed urgently, since customers will be on the opinion that they're protected against DC failures in region when they're not.