Made the description for lastIndexOf more clear
It is stated in the cases above, what happens when the string is empty or both the string and substring are empty. 

So it is not possible to have a case, where the length of the string -1 is less than zero, since the string must have at least 1 character, to enter this case. 
The two subcases are removed and the only possible case is stated instead, making it more clear for the reader, what will happen.