Update resource name limit for Policyassignments
As per the documentation, we have highlighted that the resource naming limit for policyassignment resource name can be between 1-64 characters. 

The resource name for the policyassignment resource when assigned at Management group scope is limited to 1-24 characters like what we have explained for policySetDefinition resource in the documentation. 

The limit for policyassignments resource name at Management group scope should be added to the documentation