Update virtual-networks-faq.md
Wanted to add in Considerations that ICMP traffic is not serviced by Service Endpoints. Azure customers open Support Requests after enabling Service Endpoints because they believe their routing is misconfigured after a simple tracert. This detail can lengthen troubleshooting efforts for normal performance/latency issues because the route being worked on is the incorrect one.
Originally added on to virtual-network-service-endpoints-overview.md in Considerations section but asked to send to FAQ.