Update cross-region-replication-introduction.md
IHAC customer who requested we clearly state that the destination volume is read-only until failover in our documentation. I realize it's on the fail-over page, but it may be good on this page too as this explains CRR versus the how-to. 
 
"Note: * The Azure NetApp Files destination volume is read-only until you fail-over to your destination region enabling the destination ANF volume for read and write. "