Update: Identify legacy authentication use
It should be clarified to also look during legacy authentication discovery on the **non-interactive sign-ins** page. 
There are cases where for instance POP connections stay alive after the initial interactive sign-in for many days/weeks, and don't show up again in interactive sign-in reports, even though there are hundreds of daily **non-interactive** sign-ins for the same POP connection. Given that the interactive sign-in report only goes back for 30 days, the importance of these legacy auth connections that appear only sporadically in the interactive sign-in logs could be easily missed.