mention of 443 creates a confusion with customer
TCP Port 443 is used only for HTTP communications. Either we should remove this mention of 443 is not per standards or else we should put a disclaimer that this port on the application server should not be utilized by any other service then they can use. We had a customer coming with receive connector configuration with 443 and because 443 is used by IIS for Exchange Webservices, frontend transport service fails to start. Even though you mention that this is non-standard they keep coming back with this article mentioning for SMTP you can use 443. Need to fix this.