Update search-get-started-python.md
Removing the property "value" from the documents declaration as that is failing in Python with the error “() The request is invalid. Details: parameters : A resource without a type name was found, but no expected type was specified. To allow entries without type information, the expected type must also be specified when the model is specified.”