Customer feedback added back in
One of the PMs wanted me to stress that this request is not for customer feedback but instead for feedback related to deprication.
articles/defender-for-iot/edge-security-module-deprecation.md
- Lines 34-38: Remove the section on request for feedback.
He claims he copied the style for it from here https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-tls-deprecating-1-0-and-1-1#customer-feedback.

I have added it back in and it should not be a locking issue