Added Managed Identities as a best practice
I've also clarified better the sentence about the privileged accounts, because it did seem to refer to Azure AD accounts. As a result, it looked like we are recommending not to use Azure AD and instead favor SQL credentials, which would not be correct.
Disclosure: I'm a MS FTE (simonec).