Update Durable Functions unit testing documentation
Much of the documentation for the Durable Functions unit testing
document still used obsolete types from Durable 1.x. This lead to a
very disjointed experience, both because we are discouraging customers
from using Durable 1.x, as well as the fact that we were pulling in
code samples from Durable 2.x in other places.

In addition to updating the types, some language was changed and/or
added to reflect the differences of mocking interfaces vs abstract
classes.