Update monitor-blob-storage.md
When streaming the logs to storage account, it doesn't provide any option to set retention period as to for how much time the logs should be maintained. This option is however available in storage analytics hence proposing this as a pointer for differentiation
https://docs.microsoft.com/en-us/rest/api/storageservices/Setting-a-Storage-Analytics-Data-Retention-Policy

Also, the diagnostic settings usually provides a retention period when storing diagnostics logs into storage account but this doesn't seem to be the case with Diagnostic settings preview for storage account. Hence, I have made use of word "currently" in case this is to be a future work item.
https://docs.microsoft.com/en-us/azure/azure-monitor/platform/diagnostic-settings