Update automation-windows-hrw-install.md
I wondered for a while what you meant by "run runbooks directly on the Azure or non-Azure machine". 
THE Azure or non-Azure machine? So there is only one machine? THE machine, and we can run runbooks on that particular machine? Which one, where is it?
After some time I realised perhaps you meant "an". Run runbooks directly on an Azure or non-Azure machine. Any machine, not "the" machine.
Maybe I'm wrong.