Fix Indentation Error in "Purge CDN endpoint" Sample Actions in GitHub Actions Static Site Deployment Guide
Fixed indentation errors in the sample workflow code provided in the "Set up a GitHub Actions workflow to deploy your static website in Azure Storage" guide that prevent the "Purge CDN endpoint" actions from running successfully when placeholder values are replaced with "real" values. Specifically, added an additional level of indentation to the Azure CLI commands on lines 130 and 164 to correct these errors.