Update sql-agent-extension-manually-register-single-vm.md
The Location parameter now seems to be a required parameter for New-AzSqlVM.  Otherwise you'll get prompted for Location when you run the command.  I think the sample should include the Location parameter for two reasons.  One, because it's required.  And Two, because Location is essentially what we typically refer to as the Azure region. It took a while to figure out what Location was asking for here. If we show in the example a region like "westus2" or "usgovvirginia", then it will make more sense.