No info provided about the Private IP if exists
In the above context, it mentions about the NIC but not about the Private IP which confuses the customer to proceed to move the VM to a new Availability set or not. As the customer might need the VM along with the private IP when moving to a new Availability set, more clarification helps as my customers(Azure Support Chat Team's Cxs) were confused with no clue.