Missing Prereqs
Prereq for additional subscription level permissions is not listed in the tutorial, but is listed in the prereqs linked here, https://docs.microsoft.com/en-us/azure/dms/pre-reqs. Adding the subscription level permissions here to complete the article since the other article is not mentioned at all in this tutorial. This same block is needed in several other tutorials as well and I can add those if this is accepted.  Another option might be to get rid of the prereqs in the tutorials and link to the prereqs doc directly in each tutorial, happy to make that change as well.