Add links to GitHub samples repos in quickstarts
Add links in the Azure Cache for Redis quickstart articles to the
corresponding code locations in the Azure Samples GitHub organization.

This change is inspired by the Azure Cognitive Services docs, like the
[getting started with speech-to-text article][0]. Following that style,
the Azure Cache for Redis team has collected the code from the
quickstart articles into GitHub repos in the Azure Samples org, so now
the articles can link to those repos.

[0]: https://docs.microsoft.com/azure/cognitive-services/speech-service/get-started-speech-to-text