includes: Fix New SSL Enter-PSSession instructions for Linux
In commit f59311b23ce585ee1d2e82db2aab5a270d7e6290, the option `-UseSSL`
was added to the `Enter-PSSession` line for Linux. However, if added,
it returns the following error (PowerShell 7.1.2):

```
Enter-PSSession: HTTPS on Unix does not currently support CA or CN checks.
Use the PSSessionOption -SkipCACheck and -SkipCNCheck if you are certain you
trust the server you are connecting to and the network in between.
```

Therefore, in this commit the `-UseSSL` flag is removed from the Linux
instructions.

Signed-off-by: Leonid Bloch <leonidb@asocscloud.com>