Update concept-all-sign-ins.md
Before we had a non-interactive report, we included some non-interactive exchange sign ins in the interactive report so customers would have visibility. When we created the non-interactive sign ins report, we moved legacy Exchange sign-ins to the non-interactive report since they're non interactive. This confused some customers and led to an ICM here: https://portal.microsofticm.com/imp/v3/incidents/details/230478719/home

I added a note about this change that happened in Nov 2020. We should keep that in the docs for some time