Fix link to App Insights docs rendering incorrectly
This change fixes a link to a section of the App Insights documentation which was displaying incorrectly. This file is under the Azure Batch documentation.

An extra space was added in the URL which made it invalid and also broke the markdown formatting. To fix the issue, this extra space was removed. The URL should now link and render correctly (and does in the GitHub preview).