Update tutorial-copy-data-tool.md
Added a note on  SQL 'allowing all azure services' as this has caught us out, as we originally didn't realise this would open access to all azure resources not just those in our tenant / subscription. When we did this years ago. Also added a link to private links now they are becoming available.