Add Custom Payload Detail
Added additional verbiage around the fact that if you want a custom json payload in your webhook, you cannot use the common alert schema as it will override your custom one. Would be nice to actually see a warning in the UI when you check the box for a custom payload. I just set this up and felt this was not clear enough in the documentation.