Adding additional query
I debated creating a separate article for this, but because I referenced this article to make the query I have added I felt it was best to add the query here.  Overall, if a customer wanted to generate a report of the last login attempt from all of their users, there is currently no document that handles this scenario.  I am adding the scenario to this document, because it allows that scenario to be public for customers who need a report of all users.