FAQ: Update 'How do I determine if a directory is approaching the limit size?'
Added additional clarification:

"This number translates to approximately 4 million files maximum for a 320-MB directory. However, the actual number of maximum files might be lower, depending on factors such as the number of files containing non-ASCII characters in the directory. As such, you should use the `stat` command as follows to determine whether your directory is approaching its limit. "