Help keep advice in context
As part of an internal conversation around https://github.com/microsoft/PowerShell-DSC-for-Linux/issues/764 it was suggested that we use python2 on RHEL8 since it was "recommended by Microsoft" in this document. This seems to be a misunderstanding of the context of this fix and goes against the strong preference for python3 in RHEL8. Help clarify that with this trivial improvement.