Add ASP.NET Core quickstart
This sample combines pieces of the .NET Core quickstart with pieces of
the ASP.NET quickstart and adjusts them for the ASP.NET Core context.

To stay consistent with the theme of using cross-platform command line
steps in .NET Core apps, this quickstart skips the section about using
Visual Studio to deploy to Azure and instead links to other pages about
command line steps to deploy to Azure.