Fix incorrect use of standard terminology
- replace the incorrect use of a standard term (_"Azure Cosmos DB connection string"_) with the correct term for this document's usage, i.e. **"SQL connection string""**
- use the term **"SQL connection string for Azure Cosmos DB"** to distinguish the particular type of SQL connection strings used in this document from other SQL connection strings used by Azure Synapse Analytics
- show the syntax of connection strings that, by standard, are referred to as _"Azure Cosmos DB connection strings"_
- shows what fields should map into the format of a **"SQL connection string for Azure Cosmos DB"**
- add information that the `region` field of the **"SQL connection string for Azure Cosmos DB"** is optional, and describe the default behavior

The term _"Azure Cosmos DB connection string"_ has a standard meaning for users of Azure Cosmos DB. The usage of that term in this document is incorrect and confusing. In this document, the standard term is being used to mean a **"SQL connection string"**. Although the document does provide the syntax for the SQL connection string, the document does not provide enough cues to let the reader know that what this document calls an _"Azure Cosmos DB connection string"_  is not an actual _**"Azure Cosmos DB connection string"**_. In addition, the nature of these both being connection strings means that a casual reader might not _notice the subtle differences_.

One reason the current use of the term _"Azure Cosmos DB connection string"_ is particularly problematic is that all the other documents that are part of this series that explains how to combine Azure Cosmos DB with Azure Synapse Analytics use the _correct meaning for that term_. Only in this document is the term being used to mean something different, something non-standard.