Update storage-lifecycle-management-concepts.md
The prefix match in case we want to match all blobs under `https://myaccount.blob.core.windows.net/container1/foo/...`  should be container1/foo/. There should be an extra '/' at the end as well. Otherwise, it will rather match all the blobs under container1 starting with foo and not just the ones that are under foo.