Update maintenance-and-updates.md
We are getting a lot of requests and inquiries from customers on why they aren't see live migration information in the azure portal for events that do not require a reboot. Adding in a note, will help us with customer questions on these types of issues. If there is a why, then I think we should add that as well but at least we have the public documentation updated to show customers what they should and shouldn't see in the portal.