Fixing (109) login failed error
When a user tries to sign in to Saba after 7200 seconds, if they are not required to re-enter their password in login.microsoftonline.com, they will get an error that says (109) Login failed.  This error is not well documented, but is referring to the value of the AuthnInstant attribute on the AuthnStatement line of the SAMLResponse Azure AD sends back.

Microsoft's max session lifetime appears to be 90 days (see this article https://docs.microsoft.com/en-us/azure/active-directory/conditional-access/howto-conditional-access-session-lifetime ) while Saba defaults to 7200 seconds.  If this value is not updated, users will have to clear their cache after 7200 seconds after their last password entry in Azure AD.  The value should be set to the same value as Microsoft (usually 7776000 seconds).