Update storage-sas-overview.md
We have seen cases at times where in Customer want to generate SAS from the Azure Portal however the the button comes as Greyed out. The most common cause for this is they are not selecting the resource type at the top of the page. Post selection of appropriate resource type such as Service, Container or Objects, the button shall come as enabled. It will be good to let customers know this and the first place via documentation itself. 

Hence, proposing the changes. Please review.