Update structured-app-log.md
The original documentation guides the user to setup logging configuration in `logback.xml` or `log4j2.xml`, which will be applied for all the environments.

A more practical way is to set the configurations in files with `-spring` suffix, like `logback-spring.xml`. In this file, we can set different logging configurations based on the Spring active profile. With this, we can output human readable logs for local development, while JSON format for the structured logging when deployed to Azure Spring Cloud service.