Update networking-move-limitations.md
Par internal discussion with PG, seems that Virtual Network Gateways are supported for move. They are only not supported for move if they are associated to a Public IP Standard SKU (because the limitation is on these types of ips)... with Basic Public IPs. This current document makes someone think that the gateways will never be possible to move, because in no situation we have a gateway without a public IP (it is a requirement for creating them)... so please update this doc to make it clearer.