dpdk: update install instructions
Drop Ubuntu 16.04, as it is EOL as of this month.

Update PPA address as instructed by Canonical for Ubuntu 18.04.

Note that the mentioned kernels/distros versions are only the
minimum requirements, and that newer versions will work too.

Add a note about installing on Debian too.

Add a note about the packet dump tool, which is often requested.