Remove minor version from CentOS/RHEL in tier 2 list
For CentOS and RHEL, the latest security patches other important updates are applied to the latest minor releases, so we shouldn't pin to a specific minor version in our tier 2 list. This change updates the two entries to use only the major version. It also updates the way CentOS-7 is written, to match the official name.