Adding update on assistive text.
Per customer request adding a note here referencing company branding. Per current product logic for assistive text; If there is customized text available for the tenant (which is not currently supported), we show the customized assistive text in the password change page. If there isn't customized text available for the tenant, we check if the password type for the user if on-perm or not -If the user's password is "on-perm" or "on-perm & on cloud", we don't show any assistive text. Else we show the assistive text. 
At this time we are not supporting configuring assistive text however, it is in the roadmap. Several customers have requested this via support tickets and the feature is being tracked in a work item. Last customer requested it documented here. Resubmitting this after making edits.