Update object-replication-overview.md
This is also capped for the destination storage account as well. For e.g. if there are more than 10 containers in source storage account, then only 10 of them can be mapped to destination storage account under a single replication policy. The rest need to be mapped to another destination storage account or accounts (capped to 10). 

The destination account can be specified only once and only 10 rules (source and destination container mapping) can be created.  

Proposing these changes to be clearer in documentation.

Kindly review