Swap references to Debian containers with Ubuntu
I was reading these (very helpful) docs recently and noticed what appears to be a small error: the type mentions a Debian image, and the example implementation references an Ubuntu image. I swapped these over to both be Ubuntu (treating the code as the source of truth rather than the type).  

Please let me know if this type of PR is welcome/helpful! I can close it if not.