Bad references to constrained core sizes.
Removed bad references to constrained core sizes. Removed footnote reference to constrained core sizes under "Ev3-series" section, as no VMs under that series are capable of constrained core sizes (leaving footnote 2 in that section the same as to not confuse with the footnote 1 in next section for constrained core sizes).