Including $ next to variables in CLI section.
We need to add $ next to the variables in the CLI sections of this document. If a user pastes the variables as they are currently, the CLI won't save the variables. The CLI needs the $ to save the variables. Also, the PowerShell sections in this document already includes $ next to the variables. At minimum, we should include a note to tell users they need to include $ next to the variables so the CLI will save the variables and data.