Update storage-files-quick-create-use-windows.md
This is my first proposed change, so please bear with me if I've done something wrong. It's equally possible that the change I've suggested doesn't appear for everyone. That said, when I walked through the quickstart, I was not presented with the Windows Server 2016 Datacenter option. Instead, I had to select the Windows Server 2019 Datacenter option, and then select the "Windows Server 2016 Datacenter - Gen1" option from the "Image" drop-down. I made the update by making as few changes as possible, so the order of the fields is not accurate. If the field order should be maintained, it'll involve making another change. Thanks for you consideration!